Create teacher-startup-guide
The plain text version of the teacher startup guide, with no formatting/links. I'll definitely need to return to this at the end of this project to make sure that everything in the lessons is covered in this guide. Also something worth returning to is the troubleshooting section: that could definitely do with an expansion.